% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith2numeric.R
\name{arithmetic_to_numeric}
\alias{arithmetic_to_numeric}
\title{Convert arithmetic (character) format to numeric}
\usage{
arithmetic_to_numeric(x)
}
\arguments{
\item{x}{Character string containing arithmetic operators.}
}
\value{
The numeric value from evaluating the arithmetic expression.
}
\description{
Convert arithmetic (character) format to numeric
}
\examples{
# Return numeric values
arithmetic_to_numeric("1/3")
arithmetic_to_numeric("sqrt(2)")
arithmetic_to_numeric("log(2)")
arithmetic_to_numeric("exp((-3^2)/2)/sqrt(2*pi)")

# Return NA if containing non-arithmetic string
arithmetic_to_numeric("1/3a")
}
