% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMCPShiny-package.R
\docType{package}
\name{gMCPShiny-package}
\alias{gMCPShiny}
\alias{gMCPShiny-package}
\title{gMCPShiny: Shiny App for Graphical Multiplicity}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A 'Shiny' app for 'gMCPLite' for graphical described multiple test procedures introduced in Bretz et al. (2009) \doi{10.1002/sim.3495} and Bretz et al. (2011) \doi{10.1002/bimj.201000239}. Contains instructions of multiplicity graph and graphical testing for group sequential design, described in Maurer and Bretz (2013) \doi{10.1080/19466315.2013.807748}. Implements multiplicity graph visualization and reproducible code generation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/gMCPShiny/}
  \item \url{https://github.com/Merck/gMCPShiny}
  \item \url{https://gmcp.shinyapps.io/prod/}
  \item Report bugs at \url{https://github.com/Merck/gMCPShiny/issues}
}

}
\author{
\strong{Maintainer}: Yalin Zhu \email{yalin.zhu@merck.com} (\href{https://orcid.org/0000-0003-3830-8660}{ORCID})

Authors:
\itemize{
  \item Xuan Deng \email{xuan.deng@merck.com}
  \item Keaven Anderson \email{keaven_anderson@merck.com}
  \item Yilong Zhang \email{elong0527@gmail.com}
  \item Nan Xiao \email{nan.xiao1@merck.com} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})
}

Other contributors:
\itemize{
  \item Ivan Sagalaev (highlight.js library) [contributor, copyright holder]
  \item bslib contributors (navbarPage implementation in bslib) [contributor, copyright holder]
  \item shiny contributors (navbarPage implementation in shiny) [contributor, copyright holder]
  \item fs contributors (path_sanitize implementation in fs) [contributor, copyright holder]
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder]
}

}
\keyword{internal}
