% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{headerCard}
\alias{headerCard}
\title{Card with header}
\usage{
headerCard(title, ...)
}
\arguments{
\item{title}{Card title.}

\item{...}{List of elements to include in the body of the card.}
}
\value{
Card element with header and body.
}
\description{
Card with header (Bootstrap 5).
}
\examples{
if (interactive()) {
  headerCard("Card title", "Card body")
}
}
