% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-popover.R
\name{helpLink}
\alias{helpLink}
\title{Help link}
\usage{
helpLink(href)
}
\arguments{
\item{href}{TBA}
}
\value{
TBA
}
\description{
Help link with an icon and opens in a new window or tab.
}
\examples{
NULL
}
