% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-popover.R
\name{helpPopover}
\alias{helpPopover}
\title{Help popovers}
\usage{
helpPopover(title, content)
}
\arguments{
\item{title}{TBA}

\item{content}{TBA}
}
\value{
TBA
}
\description{
Help popovers (with question mark icon)
}
\examples{
NULL
}
