% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive-palette.R
\name{hgraph_palette}
\alias{hgraph_palette}
\title{Adaptive color palette generator for hGraph}
\usage{
hgraph_palette(
  pal_name = c("gray", "Okabe-Ito", "d3.category10", "d3.category20", "d3.category20b",
    "d3.category20c", "Teal"),
  n,
  alpha = 1
)
}
\arguments{
\item{pal_name}{Color palette name.}

\item{n}{How many different colors for the selected color palette?}

\item{alpha}{Transparency level, a real number in (0, 1].}
}
\value{
A vector of color hex values.
}
\description{
Adaptive color palette generator for hGraph using
\code{pal_adaptive()} and \code{pal_ramp()}.
}
\examples{
hgraph_palette("gray", n = 10)
hgraph_palette("Okabe-Ito", n = 15)
}
