% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-matrix-buttons.R
\name{matrixButtonGroup}
\alias{matrixButtonGroup}
\alias{addMatrixRow}
\alias{delMatrixRow}
\title{Matrix buttons}
\usage{
matrixButtonGroup(inputId)

addMatrixRow(x)

delMatrixRow(x)
}
\arguments{
\item{x}{TBA}

\item{inputID}{TBA}
}
\value{
TBA
}
\description{
Buttons to add/delete rows and reset a matrix input
}
\examples{
NULL
}
