% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-fluid-page.R
\name{navbarFluidPage}
\alias{navbarFluidPage}
\title{Create a page with a top level navigation bar within a fluid container}
\usage{
navbarFluidPage(
  title,
  ...,
  id = NULL,
  selected = NULL,
  position = c("static-top", "fixed-top", "fixed-bottom"),
  header = NULL,
  footer = NULL,
  inverse = FALSE,
  collapsible = FALSE,
  fluid = TRUE,
  theme = NULL,
  windowTitle = NA,
  lang = NULL,
  brand_image,
  brand_image_width,
  brand_image_height
)
}
\arguments{
\item{title}{The title to display in the navbar}

\item{...}{\code{\link[shiny:tabPanel]{tabPanel()}} elements to include in the page. The
\code{navbarMenu} function also accepts strings, which will be used as menu
section headers. If the string is a set of dashes like \code{"----"} a
horizontal separator will be displayed in the menu.}

\item{id}{If provided, you can use \verb{input$}\emph{\code{id}} in your
server logic to determine which of the current tabs is active. The value
will correspond to the \code{value} argument that is passed to
\code{\link[shiny:tabPanel]{tabPanel()}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{position}{Determines whether the navbar should be displayed at the top
of the page with normal scrolling behavior (\code{"static-top"}), pinned at
the top (\code{"fixed-top"}), or pinned at the bottom
(\code{"fixed-bottom"}). Note that using \code{"fixed-top"} or
\code{"fixed-bottom"} will cause the navbar to overlay your body content,
unless you add padding, e.g.: \code{tags$style(type="text/css", "body
  {padding-top: 70px;}")}}

\item{header}{Tag or list of tags to display as a common header above all
tabPanels.}

\item{footer}{Tag or list of tags to display as a common footer below all
tabPanels}

\item{inverse}{\code{TRUE} to use a dark background and light text for the
navigation bar}

\item{collapsible}{\code{TRUE} to automatically collapse the navigation
elements into a menu when the width of the browser is less than 940 pixels
(useful for viewing on smaller touchscreen device)}

\item{fluid}{\code{TRUE} to use a fluid layout. \code{FALSE} to use a fixed
layout.}

\item{theme}{One of the following:
\itemize{
\item \code{NULL} (the default), which implies a "stock" build of Bootstrap 3.
\item A \code{\link[bslib:bs_theme]{bslib::bs_theme()}} object. This can be used to replace a stock
build of Bootstrap 3 with a customized version of Bootstrap 3 or higher.
\item A character string pointing to an alternative Bootstrap stylesheet
(normally a css file within the www directory, e.g. \code{www/bootstrap.css}).
}}

\item{windowTitle}{the browser window title (as a character string). The
default value, \code{NA}, means to use any character strings that appear in
\code{title} (if none are found, the host URL of the page is displayed by
default).}

\item{lang}{ISO 639-1 language code for the HTML page, such as "en" or "ko".
This will be used as the lang in the \code{<html>} tag, as in \code{<html lang="en">}.
The default (NULL) results in an empty string.}

\item{brand_image}{Path to the brand image.}

\item{brand_image_width}{Width of the brand image in pixels.}

\item{brand_image_height}{Height of the brand image in pixels.}
}
\value{
A page with a top level navigation bar within a fluid container.
}
\description{
Slightly modified from \code{\link[shiny:navbarPage]{shiny::navbarPage()}} to support:
\itemize{
\item Navbar inside a fluid container instead of full width.
\item Brand image before title.
}
}
\examples{
library(shiny)

navbarFluidPage(
  "App Title",
  tabPanel("Plot"),
  tabPanel("Summary"),
  tabPanel("Table")
)
}
