% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-rcode.R
\name{renderRcode}
\alias{renderRcode}
\alias{rcodeOutput}
\title{Render R code with syntax highlighting}
\usage{
renderRcode(
  expr,
  env = parent.frame(),
  quoted = FALSE,
  outputArgs = list(),
  delay = 100
)

rcodeOutput(outputId)
}
\arguments{
\item{expr}{An expression to evaluate.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression. If \code{expr} is a quosure and \code{quoted} is \code{TRUE},
then \code{env} is ignored.}

\item{quoted}{If it is \code{TRUE}, then the \code{\link[=quote]{quote()}}ed value of \code{expr}
will be used when \code{expr} is evaluated. If \code{expr} is a quosure and you
would like to use its expression as a value for \code{expr}, then you must set
\code{quoted} to \code{TRUE}.}

\item{outputArgs}{List of additional arguments to pass to the
output function.}

\item{delay}{Delay in milliseconds before syntax highlighting starts.}

\item{outputId}{Output variable to read the R code from.}
}
\value{
A render function similar to \code{\link[shiny:renderPrint]{shiny::renderText()}}.
}
\description{
Renders R code with syntax highlighting using the highlight.js library.
}
\examples{
if (interactive()) {
  library("shiny")

  ui <- fluidPage(
    fluidRow(
      column(
        4,
        textAreaInput(
          "rcode_in",
          label = NULL,
          width = "100\%", height = "300px",
          value = "library('shiny')\n\nset.seed(42)\nx <- runif(10)"
        )
      ),
      column(
        8,
        rcodeOutput("rcode_out")
      )
    )
  )

  server <- function(input, output) {
    output$rcode_out <- renderRcode({
      htmltools::htmlEscape(input$rcode_in)
    })
  }

  shinyApp(ui = ui, server = server)
}
}
