% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-rmd.R
\name{renderRmd}
\alias{renderRmd}
\alias{rmdOutput}
\title{Render R Markdown with syntax highlighting}
\usage{
renderRmd(
  expr,
  env = parent.frame(),
  quoted = FALSE,
  outputArgs = list(),
  delay = 100
)

rmdOutput(outputId)
}
\arguments{
\item{expr}{An expression to evaluate.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression. If \code{expr} is a quosure and \code{quoted} is \code{TRUE},
then \code{env} is ignored.}

\item{quoted}{If it is \code{TRUE}, then the \code{\link[=quote]{quote()}}ed value of \code{expr}
will be used when \code{expr} is evaluated. If \code{expr} is a quosure and you
would like to use its expression as a value for \code{expr}, then you must set
\code{quoted} to \code{TRUE}.}

\item{outputArgs}{TBA}

\item{delay}{TBA}

\item{outputId}{Output variable to read the R Markdown from.}
}
\value{
TBA
}
\description{
Description TBA
}
\examples{
NULL
}
