% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_app}
\alias{run_app}
\title{Run Shiny app}
\usage{
run_app(port = getOption("shiny.port"))
}
\arguments{
\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port between 3000:8000, excluding ports that are blocked
by Google Chrome for being considered unsafe: 3659, 4045, 5060,
5061, 6000, 6566, 6665:6669 and 6697. Up to twenty random
ports will be tried.}
}
\value{
An object that represents the app.
Printing the object or passing it to \code{\link[shiny:runApp]{shiny::runApp()}} will run the app.
}
\description{
Run Shiny app
}
\examples{
if (interactive()) {
  gMCPShiny::run_app()
}
}
