% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize-filename.R
\name{sanitize_filename}
\alias{sanitize_filename}
\title{Sanitizes file name}
\usage{
sanitize_filename(filename, replacement = "-")
}
\arguments{
\item{filename}{Input file name.}

\item{replacement}{Replacement for illegal characters.}
}
\value{
Sanitized file name.
}
\description{
Sanitizes file name (without the extension part).
Ported and modified from \code{fs::path_sanitize()}.
}
\examples{
x <- " znul/zzz.z>z/\\\\z "
paste0(sanitize_filename(x), ".rds")
}
