% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-text-addon.R
\name{textInputAddonRight}
\alias{textInputAddonRight}
\title{Text input addon}
\usage{
textInputAddonRight(
  inputId,
  label,
  value = "",
  width = NULL,
  placeholder = NULL,
  addon = NULL
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{addon}{Addon text.}
}
\value{
Text input with addon on the right side.
}
\description{
Text input with addon on the right side.
}
\examples{
textInputAddonRight(
  "filename",
  label = "Name of the report:",
  value = "gMCPShiny",
  addon = ".Rmd",
  width = "100\%"
)
}
\references{
\url{https://getbootstrap.com/docs/5.3/forms/input-group/}
}
