% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-popover.R
\name{helpLink}
\alias{helpLink}
\title{Help link}
\usage{
helpLink(href)
}
\arguments{
\item{href}{URL to open.}
}
\value{
Link element.
}
\description{
Help link with an icon and opens in a new window or tab.
}
\examples{
helpLink("https://en.wikipedia.org/wiki/List_of_Unicode_characters")
}
