% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-popover.R
\name{helpPopover}
\alias{helpPopover}
\title{Help popovers}
\usage{
helpPopover(title, content)
}
\arguments{
\item{title}{Title of the popover.}

\item{content}{Content of the popover.}
}
\value{
Popover element.
}
\description{
Help popovers with question mark icon (Bootstrap 5).
}
\examples{
helpPopover(
  "digits",
  "Number of digits past the decimal."
)
}
