% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-matrix-buttons.R
\name{matrixButtonGroup}
\alias{matrixButtonGroup}
\alias{addMatrixRow}
\alias{delMatrixRow}
\title{Matrix buttons}
\usage{
matrixButtonGroup(inputId)

addMatrixRow(x)

delMatrixRow(x)
}
\arguments{
\item{inputId}{ID of the matrix input.}

\item{x}{Name of the matrix input to add a row to or delete a row from.}
}
\value{
A matrix button group.
}
\description{
Buttons to add/delete rows and reset a matrix input.
}
\examples{
matrixButtonGroup("matrix")
}
